requirements.txt
groq
pymupdf
fpdf2
python-dotenv
requests
openai
ebooklib
beautifulsoup4
streamlit